# milktea/skills/milktea-openapi.yaml
openapi: 3.1.0
info:
  title: MilkTea HTTP API — thiết kế một quán
  version: 2.0.0
  description: 'Hợp đồng thiết kế để triển khai; không xác nhận endpoint đã có code. Domain: domain_system.md; HTTP
    envelope: system_architecture.md; security: author_rule.md. Không mô tả STOMP trong OpenAPI. QR read-only; session
    mở khi order/CASHIER open; thanh toán linh hoạt; chỉ CASHIER close. x-allowed-account-roles chỉ ràng buộc nhánh
    xác thực tài khoản, KHÔNG phải điều kiện role cho guest. Guest được cho phép theo nhánh qrAccess/tableAccess/counterAccess
    và phạm vi resource tương ứng; luôn kiểm Domain. Tooling OpenAPI không tự enforce role hay ownership.

    Nguồn: báo cáo MilkTea cập nhật 05/10/2026, 41 UC/T01–T60. Hợp đồng thiết kế; chưa xác nhận các endpoint đã
    được triển khai. v2 bổ sung UC23 thông tin quán, UC30 ngưỡng theo vị trí, UC39 sự cố/tiêu hao pha bù; version
    BIGINT truyền chuỗi; giữ giỏ/context trong kết quả đặt bàn.'
servers:
- url: http://localhost:8080
  description: Ví dụ local; dùng SERVER_PORT và origin thực của project.
paths:
...
  /api/v1/inventory/stocks:
    get:
      operationId: listStocks
      summary: Đọc tồn KHO/BEP và cảnh báo
      description: Đọc tồn KHO/BEP và cảnh báo
      tags:
      - Inventory
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StockView'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
        description: ''
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: ''
      - name: sort
        in: query
        required: false
        schema:
          type: string
          maxLength: 255
        description: Allowlist sort field; default createdAt,desc.
      - name: location
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/StockLocation'
        description: ''
      x-allowed-account-roles:
      - KITCHEN
      - ADMIN
      x-use-cases:
      - UC35
  /api/v1/inventory/materials:
    get:
      operationId: listInventoryMaterials
      summary: Bếp/admin tra nguyên liệu khi lập chứng từ
      description: Bếp/admin tra nguyên liệu khi lập chứng từ
      tags:
      - Inventory
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MaterialView'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
        description: ''
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: ''
      - name: sort
        in: query
        required: false
        schema:
          type: string
          maxLength: 255
        description: Allowlist sort field; default createdAt,desc.
      x-allowed-account-roles:
      - KITCHEN
      - ADMIN
      x-use-cases:
      - UC29
      - UC35
  /api/v1/inventory/preparation-recipes:
    get:
      operationId: listKitchenPreparationRecipes
      summary: Bếp/admin chọn công thức sơ chế
      description: Bếp/admin chọn công thức sơ chế
      tags:
      - Inventory
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PreparationRecipeView'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
        description: ''
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: ''
      - name: sort
        in: query
        required: false
        schema:
          type: string
          maxLength: 255
        description: Allowlist sort field; default createdAt,desc.
      x-allowed-account-roles:
      - KITCHEN
      - ADMIN
      x-use-cases:
      - UC33
...
  /api/v1/inventory/movements:
    get:
      operationId: listStockMovements
      summary: Đọc lịch sử tồn
      description: Đọc lịch sử tồn
      tags:
      - Inventory
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StockMovementView'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
        description: ''
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: ''
      - name: sort
        in: query
        required: false
        schema:
          type: string
          maxLength: 255
        description: Allowlist sort field; default createdAt,desc.
      - name: materialId
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/Id'
        description: ''
      - name: location
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/StockLocation'
        description: ''
      x-allowed-account-roles:
      - KITCHEN
      - ADMIN
      x-use-cases:
      - UC35
  /api/v1/inventory/imports:
    get:
      operationId: listImports
      summary: Xem chứng từ imports
      description: Xem chứng từ imports
      tags:
      - Inventory
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InventoryDocumentView'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
        description: ''
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: ''
      - name: sort
        in: query
        required: false
        schema:
          type: string
          maxLength: 255
        description: Allowlist sort field; default createdAt,desc.
      x-allowed-account-roles:
      - KITCHEN
      - ADMIN
      x-use-cases:
      - UC31
      - UC35
    post:
      operationId: recordImport
      summary: Admin nhập nguyên liệu vào KHO
      description: Ghi phiếu+dòng+tồn KHO+lịch sử cùng transaction, lượng dương. Chỉ nguyên liệu THO, quantity dương,
        unitCost không âm đúng đơn vị; lưu nguồn nhập, giá nhập và snapshot chứng từ theo từ điển dữ liệu.
      tags:
      - Inventory
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '201':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/InventoryDocumentView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 64
        description: Unpredictable client request key; same scope+payload replay returns same resource, changed
          payload returns 409.
      x-allowed-account-roles:
      - ADMIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportRequest'
      x-use-cases:
      - UC31
  /api/v1/inventory/issues:
    get:
      operationId: listIssues
      summary: Xem chứng từ issues
      description: Xem chứng từ issues
      tags:
      - Inventory
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InventoryDocumentView'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
        description: ''
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: ''
      - name: sort
        in: query
        required: false
        schema:
          type: string
          maxLength: 255
        description: Allowlist sort field; default createdAt,desc.
      x-allowed-account-roles:
      - KITCHEN
      - ADMIN
      x-use-cases:
      - UC32
      - UC35
    post:
      operationId: recordIssue
      summary: Bếp xuất KHO sang BEP không cần duyệt
      description: Kiểm đủ KHO, ghi hai vị trí+lịch sử, báo admin sau commit; không approval API. Chỉ THO, quantity
        dương; SOCHE được tạo ở BEP qua mẻ, không nhận schema này để tự tăng SOCHE.
      tags:
      - Inventory
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '201':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/InventoryDocumentView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 64
        description: Unpredictable client request key; same scope+payload replay returns same resource, changed
          payload returns 409.
      x-allowed-account-roles:
      - KITCHEN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryDocumentRequest'
      x-use-cases:
      - UC32
  /api/v1/inventory/preparation-batches:
    get:
      operationId: listPreparationBatches
      summary: Xem chứng từ preparation-batches
      description: Xem chứng từ preparation-batches
      tags:
      - Inventory
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PreparationBatchView'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
        description: ''
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: ''
      - name: sort
        in: query
        required: false
        schema:
          type: string
          maxLength: 255
        description: Allowlist sort field; default createdAt,desc.
      x-allowed-account-roles:
      - KITCHEN
      - ADMIN
      x-use-cases:
      - UC33
      - UC35
    post:
      operationId: recordPreparationBatch
      summary: Bếp ghi mẻ nguyên liệu sơ chế
      description: Trừ actualInputs THO BEP, tăng SOCHE actualYield, snapshot công thức, lý do lệch chuẩn; không
        trừ đầu vào lần nữa khi bán.
      tags:
      - Inventory
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '201':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/PreparationBatchView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 64
        description: Unpredictable client request key; same scope+payload replay returns same resource, changed
          payload returns 409.
      x-allowed-account-roles:
      - KITCHEN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreparationBatchRequest'
      x-use-cases:
      - UC33
  /api/v1/inventory/waste:
    get:
      operationId: listWasteRecords
      summary: Xem chứng từ waste
      description: Xem chứng từ waste
      tags:
      - Inventory
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '200':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InventoryDocumentView'
                  meta:
                    $ref: '#/components/schemas/PageMeta'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
        description: ''
      - name: size
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        description: ''
      - name: sort
        in: query
        required: false
        schema:
          type: string
          maxLength: 255
        description: Allowlist sort field; default createdAt,desc.
      x-allowed-account-roles:
      - KITCHEN
      - ADMIN
      x-use-cases:
      - UC34
      - UC35
    post:
      operationId: recordWaste
      summary: Bếp ghi hao hụt/sự cố
      description: KITCHEN hoặc ADMIN ghi lượng HAO_HUT chưa tiêu hao ở đúng location; reason bắt buộc. Server đối
        chiếu order/incident/ledger, từ chối debit trùng. Lượng đã debit ghi incidents/audit; làm lại sau complete
        dùng additional-consumptions, không cộng nguyên liệu.
      tags:
      - Inventory
      security:
      - accountCookie: []
      - bearerAuth: []
      responses:
        '201':
          description: Thành công; kiểm tra quyền và invariant phía server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    const: true
                  data:
                    $ref: '#/components/schemas/InventoryDocumentView'
                  meta:
                    type: 'null'
                additionalProperties: false
                required:
                - success
                - data
                - meta
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
      parameters:
      - name: X-CSRF-TOKEN
        in: header
        required: true
        schema:
          type: string
          maxLength: 1024
        description: Required for browser mutations. Fetch /api/v1/auth/csrf; never read account JWT.
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 64
        description: Unpredictable client request key; same scope+payload replay returns same resource, changed
          payload returns 409.
      x-allowed-account-roles:
      - KITCHEN
      - ADMIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WasteRequest'
      x-use-cases:
      - UC34
...
components:
  securitySchemes:
    accountCookie:
      type: apiKey
      in: cookie
      name: milktea_access_token
      description: JWT account HttpOnly; role/state/version enforced server-side.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tableAccess:
      type: apiKey
      in: header
      name: X-Table-Token
      description: Opaque access scoped to one currently OPEN TableSession, not an account role.
    counterAccess:
      type: apiKey
      in: header
      name: X-Counter-Token
      description: Opaque access scoped to one counter order; client keeps only in current page RAM.
    qrAccess:
      type: apiKey
      in: header
      name: X-QR-Code
      description: Opaque QR code identifies a table; never a staff credential.
  schemas:
    Id:
      type: string
      pattern: ^[1-9][0-9]*$
      example: '5'
    Decimal:
      type: string
      pattern: ^[0-9]+(\.[0-9]{1,3})?$
      example: '35000'
...
    MaterialType:
      type: string
      enum:
      - THO
      - SOCHE
    StockLocation:
      type: string
      enum:
      - KHO
      - BEP
    PageMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 0
        totalElements:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
      additionalProperties: false
      required:
      - page
      - size
      - totalElements
      - totalPages
...
    MaterialView:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        name:
          type: string
          maxLength: 120
        type:
          $ref: '#/components/schemas/MaterialType'
        unit:
          type: string
          maxLength: 30
        active:
          type: boolean
      additionalProperties: false
      required:
      - id
      - name
      - type
      - unit
      - active
      description: Thông tin nguyên liệu THO/SOCHE và đơn vị; ngưỡng thuộc Stock theo vị trí, không thuộc Material.
    StockView:
      type: object
      properties:
        materialId:
          $ref: '#/components/schemas/Id'
        location:
          $ref: '#/components/schemas/StockLocation'
        quantity:
          $ref: '#/components/schemas/Decimal'
        unit:
          type: string
          maxLength: 30
        threshold:
          $ref: '#/components/schemas/Decimal'
        low:
          type: boolean
      additionalProperties: false
      required:
      - materialId
      - location
      - quantity
      - unit
      - threshold
      - low
...
    PreparationRecipeView:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        name:
          type: string
          maxLength: 120
        outputMaterialId:
          $ref: '#/components/schemas/Id'
        standardYield:
          $ref: '#/components/schemas/Decimal'
        inputs:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/IngredientInput'
        active:
          type: boolean
      additionalProperties: false
      required:
      - id
      - name
      - outputMaterialId
      - standardYield
      - inputs
      - active
...
  responses:
    Error400:
      description: HTTP 400; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error401:
      description: HTTP 401; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error403:
      description: HTTP 403; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error404:
      description: HTTP 404; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error409:
      description: HTTP 409; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Error500:
      description: HTTP 500; error code/shape theo system_architecture.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
- accountCookie: []
- bearerAuth: []
...